- 로직 수정으로 컴포넌트 병합 후 비어진 컴포넌트는 삭제 처리 / 혜택 추천 모달 수정 중
