+++ src/assets/css/benefitCluster/Modal.css
@@ -4,12 +4,12 @@
     font-weight: bold;
     padding-top: 10px;
     margin-bottom: 10px;
+    margin-bottom: 20px;
 }
 
 /* common으로 만들지 고민중 */
 .benefitFilterWrap {
     padding-bottom: 17px;
-    border-bottom: 1px solid #dcdcdc;
 }
 .benefitFilterWrap .filterLine{
     display: flex;
@@ -32,4 +32,6 @@
 .benefitSearchResult {
     display: flex;
     flex-wrap: wrap;
+    border-top: 1px solid #dcdcdc;
+    padding-top: 15px;
 }
