+++ src/views/admin/benefit/components/BenefitRecommendResult.js
@@ -2,34 +2,11 @@ import { useEffect, useState } from 'react';
 import { DragDropContext } from 'react-beautiful-dnd';
 import Column from './Column';
 //import { styled } from './stitches.config';
-import { Box, Flex } from '@chakra-ui/react';
+import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import { Fragment } from 'react';
 import { styled } from "styled-components";
-
-const StyledColumns = styled('div', {
-    display: 'grid',
-    gridTemplateColumns: '1fr 2fr',
-    //   margin: "1vh auto",
-    //   width: "80%",
-    height: '100%',
-    width: '50%',
-    gap: '8px',
-});
-const Styledleff = styled('div', {
-    height: '100rem',
-
-    width: '0.005rem',
-    border: '0.01rem solid',
-    borderColor: '#000000',
-    margin: '0.5rem',
-});
-// const DynamicStyledColumns = styled("div", {
-//   display: "grid",
-//   gridTemplateColumns: "1fr 1fr 1fr",
-//   margin: "1vh auto",
-//   width: "80%",
-//   gap: "8px",
-// });
+import NFTBenefit from './NFTBenefit';
+import axios from 'axios';
 
 const DragNDropWrapper = styled.div`
     width: 50%;
@@ -42,14 +19,47 @@ const DragNDropWrapper = styled.div`
       grid-template-columns: 1fr 1fr;
     }
     & .cateBasketWrap {
+        position: re
       & .cateUnitBody {
         
       }
     }
 `;
 
-function BenefitRecommendResult({ data, setNewCombination }) {
+const CardUnitWrap = styled.div`
+    width: 50%;
+    & .subTitle.cardTit {
+        display: flex;
+        justify-content: space-between;
+        & .btnValChk {
+          font-size: 14px;
+          height: 38px;
+          border-radius: 15px;
+        }
+        & p {
+            color: #e52c2c;
+            text-align: right;
+            margin: 10px 15px 0 0;
+        }
+    }
+    & .cardUnitBody {
+        padding: 10px 7px;
+        & .cardUnit {
+            border: 1px solid #ddd;
+            border-radius: 10px;
+            & + .cardUnit {
+                margin-top: 15px;
+            }
+        }
+    }
+
+`;
+
+function BenefitRecommendResult({ data }) {
     const [columns, setColumns] = useState({});
+    const [newCombination, setNewCombination] = useState([]);
+    const [combival, setCombival] = useState(0);
+    const [curRankVal, SetCurRankVal] = useState(0);
 
     useEffect(() => {
         const newData = {};
@@ -70,6 +80,39 @@ function BenefitRecommendResult({ data, setNewCombination }) {
         setColumns(newData);
         setNewCombination([]);
     }, [data]);
+
+    useEffect(() => {
+      console.log('조합배열: Combination에서 호출');
+      console.log(newCombination);
+      let combiSum = 0;
+
+      newCombination.forEach((data, i) => {
+          let curCombiIdx = data.card_benefit_info.length;
+          let curCombiSum = 0;
+          data.card_benefit_info.forEach((d, i2) => {
+              curCombiSum += d.expectedBenefitValue;
+          });
+          combiSum += curCombiSum / curCombiIdx;
+      });
+
+      setCombival(combiSum.toFixed(0));
+  }, [newCombination]);
+
+  const handleSecondAction = () => {
+      axios({
+          url: '/benefitCluster/benefitCombination',
+          method: 'post',
+          data: { combival: combival },
+      })
+          .then((res) => {
+              console.log(res.data);
+              SetCurRankVal(res.data[res.data.length - 1].cur_rank_val);
+          })
+          .catch((err) => {
+              console.log(err);
+          });
+  };
+
 
     const onDragEnd = (result) => {
         const { source, destination } = result;
@@ -160,57 +203,65 @@ function BenefitRecommendResult({ data, setNewCombination }) {
     };
 
     return (
-        <>
-            <DragDropContext onDragEnd={onDragEnd}>
-                <DragNDropWrapper>
-                  <Box className="subTitle cateTit">Category</Box>
-                  <Box className="cateBody">
-                    <Box className="cateWrap cateStaticWrap">
-                      {Object.values(columns)
-                          .filter((col) => col.id !== "신규 조합")
-                          .map((col) => (
-                            <Column col={col} key={col.id} />
-                      ))}
-                    </Box>
-                    <Box className="cateWrap cateBasketWrap">
-                        {/* 신규조합 상자는 크기 제한이 없었으면 좋겠다. 최소 150이상 부터 가능하도록 */}
-                        {Object.values(columns)
-                            .filter((col) => col.id === "신규 조합")
-                            .map((col) => (
-                              <Fragment key={col.id}>
-                                <Column col={col} key={col.id} />
-                              </Fragment>
-                        ))}
-                    </Box>
+      <>
+        <DragDropContext onDragEnd={onDragEnd}>
+          <DragNDropWrapper>
+            <Box className="subTitle cateTit">Category</Box>
+            <Box className="cateBody">
+              <Box className="cateWrap cateStaticWrap">
+                {Object.values(columns)
+                  .filter((col) => col.id !== "신규 조합")
+                  .map((col) => (
+                    <Column col={col} key={col.id} />
+                  ))}
+              </Box>
+              <Box className="cateWrap cateBasketWrap">
+                {Object.values(columns)
+                  .filter((col) => col.id === "신규 조합")
+                  .map((col) => (
+                    <Fragment key={col.id}>
+                      <Column col={col} key={col.id} combival={combival} />
+                    </Fragment>
+                  ))}
+              </Box>
+            </Box>
+          </DragNDropWrapper>
+        </DragDropContext>
+        {newCombination && Array.isArray(newCombination) && (
+          <>
+            <CardUnitWrap>
+              <Box className="subTitle cardTit">Card Recommended
+                  {newCombination && combival > 0 && (
+                      <Box 
+                      ml="auto">
+                        <Button
+                          className="btnValChk"
+                          colorScheme="facebook"
+                          onClick={handleSecondAction}
+                        >
+                          기존 카드와 가치 평가
+                        </Button>
+                        {curRankVal !== 0 && <Text>{curRankVal} 위</Text>}
+                      </Box>
+                  )}
+              </Box>
+              <Box className="cardUnitBody">
+                {newCombination.map((benefit, index) => (
+                  <Box className="cardUnit">
+                    <NFTBenefit
+                      data={benefit}
+                      benefit_pct={benefit.benefit_pct}
+                      benefit_detail={benefit.benefit_detail}
+                    />
                   </Box>
-                </DragNDropWrapper>
-            </DragDropContext>
-        </>
+                ))}
+              </Box>
+            </CardUnitWrap>
+            
+          </>
+        )}
+      </>
     );
 }
 
 export default BenefitRecommendResult;
-
-{/* 이전코드
-            <DragDropContext onDragEnd={onDragEnd}>
-                <Flex>
-                    <StyledColumns>
-                        {Object.values(columns)
-                            .filter((col) => col.id !== '신규 조합')
-                            .map((col) => (
-                                <Column col={col} key={col.id} />
-                            ))}
-                    </StyledColumns>
-                    <Styledleff />
-                    <Flex minHeight={150} maxHeight={300} position="fixed" minH="100%" left={1000}>
-                        {Object.values(columns)
-                          .filter((col) => col.id === '신규 조합')
-                          .map((col) => (
-                              <Fragment key={col.id}>
-                                  <Column col={col} key={col.id} />
-                              </Fragment>
-                          ))}
-                  </Flex>
-              </Flex>
-          </DragDropContext>
-*/}
