+++ src/views/admin/benefit/components/Item.js
@@ -3,23 +3,6 @@ import { Draggable } from "react-beautiful-dnd";
 import { Text } from "@chakra-ui/react";
 import styled from "styled-components";
 
-/* 
-const StyledItem = styled("div", {
-  backgroundColor: "#eee",
-  borderRadius: 4,
-  padding: "4px 8px",
-  transition: "background-color .8s ease-out",
-  marginTop: 8,
-  fontWeight: 700,
-  border:
-
-  ":hover": {
-    backgroundColor: "#fff",
-    transition: "background-color .1s ease-in",
-  },
-});
-*/
-
 const StyledItem = styled.div`
     background-color: #fff;
     border: 1px solid #ccc;
@@ -41,15 +24,7 @@ const StyledTooltip = styled.div`
   margin-top: 5px;
 
 `;
-/*
-const StyledTooltip = styled("div", {
-  fontSize: "0.8em", // 툴팁 텍스트의 폰트 크기를 줄입니다.
-  backgroundColor: "#fff", // 툴팁의 배경색을 회색으로 설정합니다.
-  borderRadius: 4,
-  marginTop: 1,
-  fontWeight: 500,
-});
- */
+
 const formatNumber = (number) => {
   if (number > 0) {
     return new Intl.NumberFormat("ko-KR").format(number);
